Add the ability to add metrics based on certain scheduler events
We're currently integration our solutions with DataDog. We wanted to
integrate that ability to send metrics to that service from our
Pipeline.

Doing such, will allow us to monitor the status of our Pipeline by
looking at statistics based on metrics sent by the pipeline.

At this moment, there is only one event that's supported but as the
feature progress forward, it's easy to see that we could support a bunch
more.

I had an implementation that was fairly basic at first but after
navigating on the existing PR against the official Luigi repo I've
discovered that there an ongoing implementation of exactly what we were
trying to achieve but with similar service. Thanks to chrispalmer, I've
been able to re-use his original work to implement ours.

https://github.com/spotify/luigi/pull/2044