Fix duplicative event logging for done tasks
Fix issue were already completed task would trigger another event.

We're hooking ourselves at a point where if the task was completed in
previous time, it would still call our DataDog tracking event. When that
case happen, we don't want to log an event again because we've already
logged it.