Add support for default environment flag
Before adding this feature, we would have the metric namespaced with
different values. For instance we would have `luigi_production` for
production metrics and `luigi_staging` for staging metrics. This caused
a whole lot of problems.

Having this new feature will allow us to set up monitoring on multiple
different environment in parallel thus effectively reducing the number
of metrics we have to manage.