Change the DD tag from env to environment
In our current implementation of this pipeline, we are already sending
the `environment` parameter in all of our task.

In our DD contrib, we log all the parameters that are passed to a task
as a tag. That said, the environment tag is already set in our case.

If we were to have `env` then we would have both tags sent to DD for
every metrics / events which duplicated the numbers of that that we
trully want.

DD is clever enough that if a tag already exists, then it won't
duplicate it, so by default we want to log the `environment` so that if
it doesn't exists, it will create it, else it will use the one that we
pass to the task itself.

[ch9704]